% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calacuateRatio}
\alias{calacuateRatio}
\title{Calculate Ratios for mif Data Frame.}
\usage{
calacuateRatio(
  data,
  numerators,
  denominator,
  newUnit = "1",
  conversionFactor = 1
)
}
\arguments{
\item{data}{A mif data frame with columns \code{variable}, \code{model},
\code{scenario}, \code{region}, \code{period}.}
}
\value{
A mif data frame with changed values and new unit.
}
\description{
Changes the value of variables given in \code{numerators} by dividing by
\code{denominator} and multiplying \code{conversionFactor}. Sets \code{unit}
of these variables to \code{newUnit}.
}
