% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkGlobalOptionsProvided}
\alias{checkGlobalOptionsProvided}
\title{Check Whether Global Options Are Provided}
\usage{
checkGlobalOptionsProvided(optNames, envir = rlang::caller_env())
}
\arguments{
\item{optNames}{A character vector. The names of the variables.}

\item{envir}{The environment where to look for the variables.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
The function checks whether a variable of the name \code{optNames} is
available. If not a warning message is produced, indicating how this variable
can be set using \code{options()}.
}
