% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAreaAndBarPlots.R
\name{createAreaAndBarPlots}
\alias{createAreaAndBarPlots}
\title{Create Area and Bar Plots}
\usage{
createAreaAndBarPlots(
  data,
  vars,
  tot = NULL,
  fill = FALSE,
  orderVars = c("mean", "user", "userRev"),
  mainReg = getOption("mip.mainReg"),
  yearsBarPlot = getOption("mip.yearsBarPlot"),
  scales = "free_y"
)
}
\arguments{
\item{data}{A quitte object or an object that can be transformed into a
quitte object.}

\item{vars}{A character vector. The variables to be plotted.}

\item{tot}{A single string. A total value to be shown in the area plots.}

\item{fill}{Logical. Should the vars be normalized so that their values add
to 1? (Plot shares instead of absolute values.)}

\item{orderVars}{In what order should the variables be shown?
\describe{
  \item{\code{"mean"}}{
    Order by mean value
    (largest mean at bottom).
    The default.}
  \item{\code{"user"}}{
    As supplied by the user
    (first element of \code{vars} at top).}
  \item{\code{"userRev"}}{
    Reverse order of the one supplied by the user
    (first element of \code{vars} at bottom).}
}}

\item{mainReg}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}

\item{yearsBarPlot}{A numeric vector. The years shown in the bar plots. Use
\code{options(mip.yearsBarPlot=<value>)} to set globally.}

\item{scales}{adjusts how axes are harmonized. Default is free_y}
}
\value{
A list of plots
}
\description{
Creates the area and bar plots for layoutAreaAndBarPlots.
}
