% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAreaAndBarPlotsPlus.R
\name{createAreaAndBarPlotsPlus}
\alias{createAreaAndBarPlotsPlus}
\title{Create Area and Bar Plus Plots}
\usage{
createAreaAndBarPlotsPlus(
  data,
  tot,
  plusNum = 1,
  fill = FALSE,
  mainReg = getOption("mip.mainReg"),
  yearsBarPlot = getOption("mip.yearsBarPlot"),
  scales = "free_y"
)
}
\arguments{
\item{data}{A quitte object or an object that can be transformed into a
quitte object.}

\item{tot}{A single string. A total value to be shown in the area plots.}

\item{plusNum}{A single number. Number of "+"symbols for disaggregation.}

\item{fill}{Logical. Should the vars be normalized so that their values add
to 1? (Plot shares instead of absolute values.)}

\item{mainReg}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}

\item{yearsBarPlot}{A numeric vector. The years shown in the bar plots. Use
\code{options(mip.yearsBarPlot=<value>)} to set globally.}

\item{scales}{adjusts how axes are harmonized. Default is free_y}
}
\description{
Creates the plots for \code{\link{showAreaAndBarPlotsPlus}}
}
