% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLinePlotsWithTarget.R
\name{createLinePlotsWithTarget}
\alias{createLinePlotsWithTarget}
\title{Create Line Plots with Target}
\usage{
createLinePlotsWithTarget(data, vars, scales = "free_y", color.dim.name = NULL)
}
\arguments{
\item{data}{A quitte object or an object that can be transformed into a
quitte object.}

\item{vars}{A character vector. Usually just a single string. The variables
to be plotted.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{color.dim.name}{name for the color-dimension used in the legend}
}
\description{
Creates the plots for showLinePlotsWithTarget
}
