% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMultiLinePlots.R
\name{createMultiLinePlots}
\alias{createMultiLinePlots}
\title{Create Multi Line Plots}
\usage{
createMultiLinePlots(
  data,
  vars,
  scales = "free_y",
  nrowNum = 1,
  mainReg = getOption("mip.mainReg")
)
}
\arguments{
\item{data}{A quitte object or an object that can be transformed into a
quitte object.}

\item{vars}{A character vector. The variables to be plotted.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{nrowNum}{An integer value. Number of rows of the panel figures}

\item{mainReg}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}
}
\description{
Creates the plots for showMultiLinePlots
}
