% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRegiLinePlots.R
\name{createRegiLinePlots}
\alias{createRegiLinePlots}
\title{Create Region Comparison Line Plots}
\usage{
createRegiLinePlots(
  data,
  vars,
  scales = "free_y",
  excludeMainRegion = TRUE,
  mainReg = getOption("mip.mainReg")
)
}
\arguments{
\item{data}{A quitte object or an object that can be transformed into a
quitte object.}

\item{vars}{A character vector. The variables to be plotted.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{excludeMainRegion}{A single logical value. Should the main region be
excluded (or shown in the same plot)?}

\item{mainReg}{A single string. Use options(mip.mainReg=<value>) to set globally.}
}
\description{
Creates the plots for showRegiLinePlots
}
