% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVariableGroups.R
\name{extractVariableGroups}
\alias{extractVariableGroups}
\title{extractVariableGroups}
\usage{
extractVariableGroups(x)
}
\arguments{
\item{x}{a vector of variable names}
}
\value{
a named list of variable groups with group name as name and vector of entities as content
}
\description{
Groups variable names by groups based on the |+| separators given in the variable names
}
\examples{
x <- c("a|+|1|+|aa","a|+|2|abc","a|+|1|+|bb","a|+|1|+|cc","a|+|3|+|aa","a|+|3|+|bb")
mip:::extractVariableGroups(x)
}
\seealso{
\code{\link{plotstyle.add}}
}
\author{
David Klein, Jan Philipp Dietrich
}
