% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLegend}
\alias{getLegend}
\title{Get Legend of a ggplot.}
\usage{
getLegend(plt)
}
\arguments{
\item{plt}{A ggplot object.}
}
\value{
The legend of \code{plt} or \code{NULL} if no legend was found.
}
\description{
The function extracts the legend of a ggplot object.
}
