% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize}
\alias{harmonize}
\title{Harmonization of model data to historical data, using harmonization methods
of aneris, ported to R.
See: https://github.com/iiasa/aneris/blob/master/aneris/methods.py}
\usage{
harmonize(
  df,
  hist,
  finalYear = "2050",
  harmonizeYear = "2015",
  method = "ratio",
  suffix = ""
)
}
\arguments{
\item{df}{data frame with model data to be harmonized}

\item{hist}{data frame with historical data t be used for harmonization}

\item{finalYear}{when should harmonized data match model data again?}

\item{harmonizeYear}{when should harmonization begin?
sets model data = reference data for this year}

\item{method}{harmonization method, currently supported methods are
"ratio" and "offset"}

\item{suffix}{to be appended to harmonized variables}
}
\description{
Harmonization of model data to historical data, using harmonization methods
of aneris, ported to R.
See: https://github.com/iiasa/aneris/blob/master/aneris/methods.py
}
\author{
Falk Benke
}
