% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identifierModelScen}
\alias{identifierModelScen}
\title{add identifier based on model and scenario names}
\usage{
identifierModelScen(x)
}
\arguments{
\item{x}{A quitte object}
}
\value{
A factor. If more than one model but only one scenario, use model.
If more than one scenario but only one model, use scenario. Else, combine them.
}
\description{
add identifier based on model and scenario names
}
