% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAreaAndBarPlots.R
\name{layoutAreaAndBarPlots}
\alias{layoutAreaAndBarPlots}
\title{Layout Area and Bar Plots
Layouts the area and bar plots created by createAreaAndBarPlots, which can
then be displayed using showAreaAndBarPlots}
\usage{
layoutAreaAndBarPlots(items)
}
\arguments{
\item{items}{list of ggplot2 objects returned by createAreaAndBarPlots, these
can be modified before passing them, but structure of the list must remain
intact for layouting to work}
}
\value{
Arranged plots
}
\description{
Layout Area and Bar Plots
Layouts the area and bar plots created by createAreaAndBarPlots, which can
then be displayed using showAreaAndBarPlots
}
