% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLinePlots.R
\name{layoutLinePlots}
\alias{layoutLinePlots}
\title{Layout Line Plots}
\usage{
layoutLinePlots(items)
}
\arguments{
\item{items}{list of ggplot2 objects returned by createAreaAndBarPlots, these
can be modified before passing them, but structure of the list must remain
intact for layouting to work}
}
\value{
Arranged plots
}
\description{
Layouts the line plots created by createLinePlots, which can then be displayed using
showLinePlots
}
