% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longestCommonPrefix}
\alias{longestCommonPrefix}
\title{Longest Common Prefix}
\usage{
longestCommonPrefix(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A single string. The longest common prefix of \code{x}.
}
\description{
Longest Common Prefix
}
