% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipConvergence.R
\name{mipConvergence}
\alias{mipConvergence}
\title{Create REMIND convergence overview}
\usage{
mipConvergence(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\description{
Create REMIND convergence overview
}
\examples{

  \dontrun{
    mipConvergence(gdx="fulldata.gdx")
  }

}
\author{
Renato Rodrigues, Falk Benke
}
