% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenToolMAgPIE.R
\name{scenToolMAgPIE}
\alias{scenToolMAgPIE}
\title{scenToolMAgPIE}
\usage{
scenToolMAgPIE(file = NULL, valfile = NULL)
}
\arguments{
\item{file}{report data. Can be a CSV/MIF file or rds file with a quitte object (saved with saveRDS). file can also be a vector of rds files. NULL by default; in this case the user can upload files directly in the tool}

\item{valfile}{validation data. Can be a CSV/MIF file or rda/RData file with a quitte object (saved with saveRDS). NULL by default; in this case the user can upload files directly in the tool}
}
\description{
scenToolMAgPIE allows to explore and visualize time series of modelling results. The app is based on shiny opens in an external web brower. For more details: https://github.com/flohump/scenTool
}
\examples{
  \dontrun{
    scenToolMAgPIE("testdata.mif")
  }

}
\author{
Florian Humpenoeder
}
