% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratchBar.R
\name{scratchBar}
\alias{scratchBar}
\title{scratchBar}
\usage{
scratchBar(x, complete = "default", simplify = "default", ...)
}
\arguments{
\item{x}{an object that can be converted to a quitte (e.g. a quitte object or a magpie objet)}

\item{complete}{"default" or list with specifications for scratchComplete}

\item{simplify}{"default" or list with specifications for scratchSimplify}

\item{...}{furhter arguments handed on to mipBarYearData function}
}
\value{
ggplot object
}
\description{
Fast visualization of a magpie object or quitte object as bar plot.
If available, the years 2020 and 2050 will be selected; if not available the first and last year of the available years.
}
\examples{

  \dontrun{
    x <- Intake(gdx)
  }


}
\author{
Benjamin Leon Bodirsky
}
