% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAreaAndBarPlots.R
\name{showAreaAndBarPlots}
\alias{showAreaAndBarPlots}
\title{Show Area and Bar Plots}
\usage{
showAreaAndBarPlots(
  data,
  vars,
  tot = NULL,
  fill = FALSE,
  mainReg = getOption("mip.mainReg"),
  yearsBarPlot = getOption("mip.yearsBarPlot")
)
}
\arguments{
\item{data}{A quitte object.}

\item{vars}{A character vector. The variables to be plotted.}

\item{tot}{A single string. A total value to be shown in the area plots.}

\item{fill}{Logical. Should the vars be normalized so that their values add
to 1? (Plot shares instead of absolute values.)}

\item{mainReg}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}

\item{yearsBarPlot}{A numeric vector. The years shown in the bar plots. Use
\code{options(mip.yearsBarPlot=<value>)} to set globally.}
}
\value{
\code{NULL} is returned invisible.
}
\description{
Creates 4 sets of plots from scenario data and shows them.
}
\details{
Creates 2 sets of area plots (main region + others) and 2 sets of bar plots
(main region + others) of the variables specified in vars over time. For area
plots, faceting is done by region and scenario; for bar plots over region. If
a variables is given in tot, this is shown as a black line. If not the sum of
the values of vars is drawn. If \code{fill=TRUE}, the values of vars are
divided by the values of tot to show share of total. The plots arranged and
shown.
}
\examples{
\dontrun{
items <- c(
  "FE|CDR",
  "FE|Transport",
  "FE|Buildings",
  "FE|Industry")
showAreaAndBarPlots(data, items)
}
}
