% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAreaAndBarPlotsPlus.R
\name{showAreaAndBarPlotsPlus}
\alias{showAreaAndBarPlotsPlus}
\title{Show Area and Bar Plots Utilizing '+'-Notation}
\usage{
showAreaAndBarPlotsPlus(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link{createAreaAndBarPlotsPlus}}
  \describe{
    \item{\code{tot}}{A single string. A total value to be shown in the area plots.}
    \item{\code{plusNum}}{A single number. Number of "+"symbols for disaggregation.}
    \item{\code{data}}{A quitte object or an object that can be transformed into a
quitte object.}
    \item{\code{fill}}{Logical. Should the vars be normalized so that their values add
to 1? (Plot shares instead of absolute values.)}
    \item{\code{mainReg}}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}
    \item{\code{yearsBarPlot}}{A numeric vector. The years shown in the bar plots. Use
\code{options(mip.yearsBarPlot=<value>)} to set globally.}
    \item{\code{scales}}{adjusts how axes are harmonized. Default is free_y}
  }}
}
\value{
\code{NULL} is returned invisible.
}
\description{
Automatically infers disaggregation of a variable using the '+'-notation and
calls \code{\link{showAreaAndBarPlots}}.
}
\details{
The function requires \code{data} to have a character column named
\code{varplus} containing variable names that use the '+'-notaion. The
function searches for values in \code{varplus} that start with \code{tot}
followed by \code{"|"}, \code{plusNum}-times \code{"+"}, and \code{"|"}.
These variables are then used in a call to \code{\link{createAreaAndBarPlots}}.
}
\examples{
\dontrun{
options(mip.yearsBarPlot = c(2010, 2030, 2050, 2100))
options(mip.mainReg = "World")
data <- as.quitte(data)
showAreaAndBarPlotsPlus(data, "SE|Liquids")
}
}
