% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLinePlots.R
\name{showLinePlots}
\alias{showLinePlots}
\title{Show Line Plots}
\usage{
showLinePlots(
  data,
  vars = NULL,
  scales = "free_y",
  mainReg = getOption("mip.mainReg")
)
}
\arguments{
\item{data}{A quitte object.}

\item{vars}{A character vector. Usually just a single string. The variables
to be plotted. If \code{NULL} all rows from \code{data} are plotted.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{mainReg}{A single string. The plots for this region are shown enlarged. Use \code{options(mip.mainReg=<value>)} to set globally.}
}
\value{
\code{NULL} is returned invisible.
}
\description{
Creates 2 sets of line plots from scenario data and shows them.
}
\details{
Two sets of line plots are shown (main region + others), depicting the values
in \code{vars} over time. Faceting is done by \code{region}. The plots
arranged and shown.
}
\examples{
\dontrun{
showLinePlots(data, "Policy Cost|GDP Loss")
}
}
