% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLinePlotsWithTarget.R
\name{showLinePlotsWithTarget}
\alias{showLinePlotsWithTarget}
\title{Show Line Plots With Target}
\usage{
showLinePlotsWithTarget(data, vars, scales = "free_y")
}
\arguments{
\item{data}{A data frame created by loading scenario and historical mifs.}

\item{vars}{A character vector. Usually just a single string. The variables
to be plotted.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}
}
\value{
\code{NULL} is returned invisible.
}
\description{
Shows line plots of a variable with additional target data.
}
\details{
Creates a line plot showing single line plot of vars over time. Additionally
target values given in variables of the form \code{<vars>|target|<sth>} are
shown. The plot is shown.
}
\examples{
\dontrun{
showLinePlotsWithTarget(data, "Emi|GHG")
}
}
