% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMultiLinePlots.R
\name{showMultiLinePlots}
\alias{showMultiLinePlots}
\title{Show Multi-Line Plots}
\usage{
showMultiLinePlots(
  data,
  vars,
  scales = "free_y",
  mainReg = getOption("mip.mainReg")
)
}
\arguments{
\item{data}{A quitte object.}

\item{vars}{A character vector. The variables to be plotted.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{mainReg}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}
}
\value{
\code{NULL} is returned invisible.
}
\description{
Show 2 sets of plots with different regions in the same plot (value over time).
}
\details{
Creates two plots (main region + others) with the values of \code{vars} over
time. Different regions are shown in the same plot. Faceting is done by
\code{variable}. The plots arranged and shown.
}
\examples{
\dontrun{
items <- c(
  "FE|Transport pCap",
  "FE|Buildings pCap",
  "FE|Industry pCap")
showMultiLinePlots(data, items)
}
}
