% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMultiLinePlotsByVariable.R
\name{showMultiLinePlotsByVariable}
\alias{showMultiLinePlotsByVariable}
\title{Show Multi-Line Plots by Variable}
\usage{
showMultiLinePlotsByVariable(
  data,
  vars,
  xVar,
  scales = "free_y",
  mainReg = getOption("mip.mainReg"),
  histRefModel = getOption("mip.histRefModel")
)
}
\arguments{
\item{data}{A data frame created by loading scenario and historical mifs.}

\item{vars}{A character vector. The variables to be plotted.}

\item{xVar}{A single string. The variable for the x-axis.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{mainReg}{A single string. The plots for this region are shown enlarged. Use \code{options(mip.mainReg=<value>)} to set globally.}

\item{histRefModel}{A named character vector identifying the unique model to be
chosen for historical data. Use \code{options(mip.histRefModel=<value>)} to
set globally.}
}
\value{
\code{NULL} is returned invisible.
\dontrun{
items <- c(
  "FE|Transport pCap",
  "FE|Buildings pCap",
  "FE|Industry pCap")
showMultiLinePlotsByVariable(data, items, "GDP|PPP pCap")
}
}
\description{
Show plots with different regions in the same plot; x-axis variable chosen by
user.
}
\details{
Same as \code{\link{showMultiLinePlots}} but with the variable specified by
\code{xVar} on x-axis. For every y-axis-value, we need a unique x-axis-value.
For historical data, there may be several sources / models of the same
variable. For the x-axis-variable a unique historical source / model is
chosen via \code{histRefModel}.
}
