% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRegiLinePlots.R
\name{showRegiLinePlots}
\alias{showRegiLinePlots}
\title{Show Line Plots for Region Comparison}
\usage{
showRegiLinePlots(
  data,
  vars,
  scales = "free_y",
  excludeMainRegion = TRUE,
  mainReg = getOption("mip.mainReg")
)
}
\arguments{
\item{data}{A quitte object.}

\item{vars}{A character vector. The variables to be plotted.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{excludeMainRegion}{A single logical value. Should the main region be
excluded (or shown in the same plot)?}

\item{mainReg}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}
}
\value{
\code{NULL} is returned invisible.
}
\description{
Shows line plots of a variable with different regions in the same plot and
faceting by scenario.
}
\details{
Creates one set of plots with the values of \code{vars} over time. Does not
show historical data. Different regions are shown in the same plot. Faceting
is done by \code{scenario}. The plots arranged and shown.
}
\examples{
\dontrun{
showRegiLinePlots(data, "Price|Carbon")
}
}
