% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sideBySidePlots.R
\name{sideBySidePlots}
\alias{sideBySidePlots}
\title{sideBySidePlots}
\usage{
sideBySidePlots(ggplots, margin = 0.05)
}
\arguments{
\item{ggplots}{A list of ggplots. These plots are converted via plotly::ggplotly, however, passing plotly plots is
not allowed, because they cannot be customized anymore. When using mipIterations set the argument
returnGgplots to TRUE ot get ggplots instead of plotly plots.}

\item{margin}{Margin between plots, passed on to plotly::subplot.}
}
\value{
A tagList containing an h3 for the titles and the ggplots in a plotly::subplot.
}
\description{
Show multiple plots side by side in one row, using the same plotly slider (frame aesthetic) if that is used.
}
\seealso{
\code{\link{mipIterations}}
}
\author{
Pascal Führlich
}
