% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mip.R
\name{theme_mip}
\alias{theme_mip}
\title{MIP theme settings}
\usage{
theme_mip(size = 12)
}
\arguments{
\item{size}{Font size}
}
\description{
MIP theme settings
}
\examples{

\dontrun{
  p <- mipArea(x) + theme_mip(10)
}
}
\author{
Jan Philipp Dietrich
}
