% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateData}
\alias{validateData}
\title{Validate mif Data Frame}
\usage{
validateData(data, removeSuperfluousCols = FALSE)
}
\arguments{
\item{data}{The object to be checked.}

\item{removeSuperfluousCols}{A single truth-value. Should unrecognized
columns be removed? Set this to \code{TRUE} to make sure that the data
frame does not contain columns that create unforeseen results when using
non-standard evaluation in \code{dplyr}-functions.}
}
\value{
Returns \code{data} (with less columns if required).
}
\description{
Checks whether \code{data} is a data frame and has the right columns to be a
mif data frame. Throws error if not ok.
}
