% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationpdf.R
\name{validationpdf}
\alias{validationpdf}
\title{Create a validation PDF file}
\usage{
validationpdf(x, hist, file = "validation.pdf", style = "comparison",
  only_historical = FALSE, digits = 3, filter = NULL, prefix = NULL,
  hideEmptySection = FALSE, show_stats = TRUE, debug = getOption("debug"))
}
\arguments{
\item{x}{Data to be validated. All formats allowed which can be converted to quitte (including characters containing the path to a mif file)}

\item{hist}{Validation data.All formats allowed which can be converted to quitte (including characters containing the path to a mif file)}

\item{file}{file name of the output PDF or a Sweave object. If a sweave object is provided the function will return the updated object, otherwise
it will write its content to the file}

\item{style}{data style for the returned data. Currently available: "trafficlight", "detailed", "comparison"}

\item{only_historical}{boolean deciding whether only historical data should be used for validation or also projections from other sources}

\item{digits}{integer indicating the number of digits to be shown.}

\item{filter}{Additional filter to be applied on the data to only plot a subset of the provided data}

\item{prefix}{Prefix which will be put in front of each part title (useful if validation is integrated into a bigger document)}

\item{hideEmptySection}{removes sections in output file which would be empty for the reason that variables in input 'x' has have correspondance in the hist file}

\item{show_stats}{boolean specifying whether additional statistic section should show up or not}

\item{debug}{Switch to activate or deactivate debug mode.}
}
\description{
Create a validation PDF file
}
\author{
Jan Philipp Dietrich
}
