% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warnMissingVars}
\alias{warnMissingVars}
\title{Warn If Variables Are Missing}
\usage{
warnMissingVars(data, vars)
}
\arguments{
\item{data}{A data frame with a column named \code{variable}.}

\item{vars}{A character vector.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Generates a warning if some of the variable names in \code{vars} vars are not
entries of the variables-column of \code{data}.
}
