% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calacuateRatio}
\alias{calacuateRatio}
\title{Calculate Ratios for Quitte Objects.}
\usage{
calacuateRatio(
  data,
  numerators,
  denominator,
  newUnit = "1",
  conversionFactor = 1
)
}
\arguments{
\item{data}{A quitte object.}

\item{numerators}{A character vector. Entries in the variable column of data.}

\item{denominator}{A single string. An entry in the variable column of data.}

\item{newUnit}{A single string.}

\item{conversionFactor}{A single numerical value.}
}
\value{
A quitte object with changed values and new unit.
}
\description{
Changes the value of variables given in \code{numerators} by dividing by
\code{denominator} and multiplying \code{conversionFactor}. Sets \code{unit}
of these variables to \code{newUnit}.
}
