% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLinePlots.R
\name{createLinePlots}
\alias{createLinePlots}
\title{Create Line Plots}
\usage{
createLinePlots(
  data,
  vars = getVars(as.quitte(data)),
  histVars = vars,
  scales = "free_y",
  ylim = 0,
  show.dots = TRUE,
  color.dim.name = NULL,
  mainReg = getOption("mip.mainReg"),
  color.dim.manual = NULL,
  histModelsExclude = character(),
  target = NULL,
  vlines = NULL,
  thresholds = getOption("mip.validationThresholds")
)
}
\arguments{
\item{data}{A quitte object or an object that can be transformed into a
quitte object.}

\item{vars}{A character vector of variables to be plotted. Defaults to all
variables in `data`.}

\item{histVars}{A character vector of historical variables to be plotted.
Defaults to `vars`.}

\item{scales}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}

\item{ylim}{y limits}

\item{show.dots}{If TRUE: shows geom_point dots on line. If FALSE: only plots geom_line without geom_point}

\item{color.dim.name}{name for the color-dimension used in the legend}

\item{mainReg}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}

\item{color.dim.manual}{optional vector with manual colors replacing default
colors of color.dim, default is \code{NULL}.}

\item{histModelsExclude}{A character vector with historical models to
exclude.}

\item{target}{optional string, model variable to be plotted with dots (indicating targets)}

\item{vlines}{period used for vertical line}

\item{thresholds}{optional data.frame with validation thresholds as
returned by \code{piamValidation::getThresholdBands()}.
Thresholds are drawn as colored background bands. Defaults to
\code{getOption("mip.validationThresholds")}.}
}
\value{
List of ggplot objects
}
\description{
Creates the line plots for layoutLinePlots.
}
