% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVariableGroups.R
\name{extractVariableGroups}
\alias{extractVariableGroups}
\title{extractVariableGroups}
\usage{
extractVariableGroups(x, keepOrigNames = FALSE, sorted = FALSE)
}
\arguments{
\item{x}{a vector of variable names}

\item{keepOrigNames}{if set, the returned list contains the original variables
(to the value of which the grouped ones have to sum up) as names instead of
made up group names, if they exist.}

\item{sorted}{boolean, indicating whether the variables within each group should be returned alp}
}
\value{
a named list of variable groups with group name as name and vector of entities as content
}
\description{
Groups variable names by summation groups based on the |+| separators given in the variable names.
If no |+| are present the function will try to derive summation groups based on | separators.
}
\examples{
x <- c("a|+|1|+|aa","a|+|2|abc","a|+|1|+|bb","a|+|1|+|cc","a|+|3|+|aa","a|+|3|+|bb")
mip::extractVariableGroups(x)
}
\seealso{
\code{\link{plotstyle.add}}
}
\author{
Anastasis Giannousakis, David Klein, Jan Philipp Dietrich
}
