\name{FLAS}
\alias{FLAS}
\docType{data}
\title{
Data that was collected to investigate the usefulness of FLAS
}
\description{
Raymond (1987) describes data that were collected to investigate the usefulness of a newly developed instrument, the Foreign Language Attitude Scale (FLAS), for predicting success in the study of foreign languages. The dataset contains missing values.
}
\usage{data(FLAS)}
\format{
  A data frame with 279 observations on the following 12 variables.
  \describe{
    \item{\code{LAN}}{foreign language studied (1=french, 2=spanish, 3=german, 4=russian).}
    \item{\code{AGE}}{age group (1 = less than 20, 2 = 20+).}
    \item{\code{PRI}}{number of prior foreign language courses (1=none, 2=1-2, 3=3+).}
    \item{\code{SEX}}{1=male, 2=female.}
    \item{\code{FLAS}}{score on foreign language attitude scale.}
    \item{\code{MLAT}}{Modern Language Aptitude Test, fourth subtest score.}
    \item{\code{SATV}}{Scholastic Attitude Test, verbal score.}
    \item{\code{SATM}}{Scholastic Attitude Test, math score.}
    \item{\code{ENG}}{score on Penn State English placement exam.}
    \item{\code{HGPA}}{high scool grad point average.}
    \item{\code{CGPA}}{current college grad point average.}
    \item{\code{GRD}}{final grad in foreign language course (2=A, 1=B or lower).}
  }
}
\details{
This dataset is an adjusted version of the original dataset by Dr. Mark Raymond. The adjustments can be looked up in Schafer (1997, p.200).
}
\references{
Schafer, J.L. (1997), Analysis of Incomplete Multivariate Data. Chapman and Hall, London.
}
\keyword{Datasets}
