% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlotData.R
\name{getPlotData}
\alias{getPlotData}
\title{getPlotData}
\usage{
getPlotData(symbolName, pathToGdx = ".", ...)
}
\arguments{
\item{symbolName}{The name of a symbol to be extracted from gdx.}

\item{pathToGdx}{Path to one or more gdx files or a a path to a folder with fulldata gdx files. If multiple paths
are provided each one represents data after a specific iteration. The order of paths should match
iteration order, e.g. pathToGdx[1] should hold data for the first iteration, pathToGdx[2] for the
second iteration etc. If the path to a folder is given the fulldata gdx files in it are used.}

\item{...}{Additional arguments passed to gdxrrw::rgdx.}
}
\value{
A data frame with data from the given gdx file(s). If multiple gdx files are provided an additional
        "iteration" column is added. The iteration value will be 1 for data rows from the first gdx, 2 for the second
        etc. The last column will always be the actual value column called <symbolName>.
}
\description{
Get ready-to-plot data from one or more gdx files.
}
\seealso{
\code{\link{mipIterations}}, \code{\link{dataframeFromGdx}}
}
\author{
Pascal Führlich
}
