% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLinePlots.R
\name{layoutLinePlots}
\alias{layoutLinePlots}
\title{Layout Line Plots}
\usage{
layoutLinePlots(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=createLinePlots]{createLinePlots}}
  \describe{
    \item{\code{vars}}{A character vector of variables to be plotted. Defaults to all
variables in `data`.}
    \item{\code{histVars}}{A character vector of historical variables to be plotted.
Defaults to `vars`.}
    \item{\code{scales}}{A single string. choose either \code{"free_y"} or \code{"fixed"}.}
    \item{\code{ylim}}{y limits}
    \item{\code{show.dots}}{If TRUE: shows geom_point dots on line. If FALSE: only plots geom_line without geom_point}
    \item{\code{color.dim.name}}{name for the color-dimension used in the legend}
    \item{\code{histModelsExclude}}{A character vector with historical models to
exclude.}
    \item{\code{color.dim.manual}}{optional vector with manual colors replacing default
colors of color.dim, default is \code{NULL}.}
    \item{\code{target}}{optional string, model variable to be plotted with dots (indicating targets)}
    \item{\code{vlines}}{period used for vertical line}
    \item{\code{data}}{A quitte object or an object that can be transformed into a
quitte object.}
    \item{\code{mainReg}}{A single string. The plots for this region are shown enlarged.
Use \code{options(mip.mainReg=<value>)} to set globally.}
  }}
}
\value{
Arranged plots
}
\description{
Layouts the line plots created by createLinePlots, which can then be displayed using
showLinePlots
}
