% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip-package.R
\docType{package}
\name{mip-package}
\alias{mip-package}
\alias{_PACKAGE}
\alias{mip}
\title{The MIP R package}
\description{
Contains the routines for plotting multi model and multi scenario comparisons
}
\details{
\tabular{ll}{ Package: \tab mip\cr Type: \tab Package\cr Version: \tab
7.6\cr Date: \tab 2016-06-13\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mip}
  \item \doi{10.5281/zenodo.1158586}
  \item Report bugs at \url{https://github.com/pik-piam/mip/issues}
}

}
\author{
David Klein

Maintainer: Anastasis Giannousakis <giannou@pik-potsdam.de>
}
