% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPercentiles.R
\name{plotPercentiles}
\alias{plotPercentiles}
\title{Comparison plots show user selected variables obtained from different scenario
runs}
\usage{
plotPercentiles(df, scenarios = NULL, variables = NULL)
}
\arguments{
\item{df}{`quitte` style data frame containing all variables for each scenario}

\item{scenarios}{Character vector contains names of the desired scenarios. If none, all scenarios will be displayed}

\item{variables}{Character vector contains names of the desired variables. If none, all variables will be displayed}
}
\description{
Comparison plots show user selected variables obtained from different scenario
runs
}
\author{
Tonn Rueter
}
