% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_legend.R
\name{shorten_legend}
\alias{shorten_legend}
\title{Shorten legend names to a given length}
\usage{
shorten_legend(x, maxchar = 20, identical_only = FALSE, sep = c(" ", "|",
  "_"))
}
\arguments{
\item{x}{A character vector or a factor vector that should be shortened}

\item{maxchar}{Maximum number of characters allowed}

\item{identical_only}{If set to TRUE identical parts in the name will be removed, independent of the character length (maxchar will be ignored!)}

\item{sep}{a vector of characters which should be interpreted as separators}
}
\description{
Function will try first to return vector as it is, then remove elements which are identical between all
elements and finally cut the end of the character vectors so that it fits the given maxchar setting. Underscores
will be replaced with empty spaces for further processing
}
\examples{
a <- c("Model Scenario_BLUB", "Model Scenario_BLA", "Model Scenario_BLA_BLA_BLUB")

# do nothing
shorten_legend(a,30)
# remove identical parts
shorten_legend(a,15)
# or ...
shorten_legend(a,identical_only=TRUE)
# cutoff end of string
shorten_legend(a,5)
}
\author{
Jan Philipp Dietrich
}
