% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAreaAndBarPlotsPlus.R
\name{showAreaAndBarPlotsPlus}
\alias{showAreaAndBarPlotsPlus}
\title{Show Area and Bar Plots Utilizing '+'-Notation}
\usage{
showAreaAndBarPlotsPlus(
  data,
  tot,
  plusNum = 1,
  fill = FALSE,
  mainReg = getOption("mip.mainReg"),
  yearsBarPlot = getOption("mip.yearsBarPlot")
)
}
\arguments{
\item{data}{A data frame created by loading scenario and historical mifs.}

\item{tot}{A single string. A total value to be shown in the area plots.}

\item{plusNum}{A single number. Number of "+"symbols for disaggregation.}

\item{fill}{Logical. Should the vars be normalized so that their values add to 1? (Plot shares instead of absolute values.)}

\item{mainReg}{A single string. The plots for this region are shown enlarged. Use \code{options(mip.mainReg=<value>)} to set globally.}

\item{yearsBarPlot}{A numeric vector. The years shown in the bar plots. Use \code{options(mip.yearsBarPlot=<value>)} to set globally.}
}
\value{
\code{NULL} is returned invisible.
}
\description{
Automatically infers disaggregation of a variable using the '+'-notaion and
calls \code{\link{showAreaAndBarPlots}}.
}
\details{
The function requires \code{data} to have to column named \code{variablePlus}
containing variable names that use the '+'-notaion. The function searches for
values in \code{variablePlus} that start with \code{tot} followed by
\code{"|"}, \code{plusNum}-times \code{"+"}, and \code{"|"}. These variables
are then used in a call to \code{\link{showAreaAndBarPlots}}.
}
\examples{
\dontrun{
showAreaAndBarPlotsPlus(data, "SE|Liquids")
}
}
